pyproject.toml
[project]
name = "mcp-sqlalchemy"
version = "0.1.0"
description = "A sample project that builds MCP clients with llamacpp and runekaagaard/mcp-alchemy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.57.0",
    "mcp>=1.2.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
    "httpx>=0.24.1",
    "streamlit>=1.31.1",
    "pyodbc>=5.2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = [
  "src/mcp_openai",
]
